backend/requirements.txt
google-generativeai
psycopg2-binary
python-dotenv
pgvector


